+++ ajedrez/bishop.py
@@ -1,6 +1,6 @@
-class Queen:
+class Bishop:
     def __init__(self, color):
         self.color = color
 
     def __str__(self):
-        return "♕" if self.color == "WHITE" else "♛"
+        return "♗" if self.color == "WHITE" else "♝"