+++ test/test_horse.py
@@ -2,4 +2,4 @@ import unittest
 from ajedrez.horse import Horse
 
 class TestHorse(unittest.TestCase):
-    
+    ...